Рефакторинг пространств имён и папок
