+++ FDTD/Space1D/Boundaries/Boundary1D.cs
@@ -1,0 +1,7 @@
+﻿namespace FDTD.Space1D.Boundaries
+{
+    public abstract class Boundary1D
+    {
+        public abstract void Process(double[] Field);
+    }
+}
