+++ FDTD/Space1D/Boundaries/Boundary1DMin.cs
@@ -1,0 +1,4 @@
+﻿namespace FDTD.Space1D.Boundaries
+{
+    public abstract class Boundary1DMin : Boundary1D { }
+}
