+++ FDTD/Space2D/Boundaries/Boundary2D.cs
@@ -1,0 +1,7 @@
+﻿namespace FDTD.Space2D.Boundaries
+{
+    public abstract class Boundary2D
+    {
+        public abstract void Process(double[,] Field);
+    }
+}
