+++ FDTD/Space2D/Boundaries/Boundary2DMaxX.cs
@@ -1,0 +1,4 @@
+﻿namespace FDTD.Space2D.Boundaries
+{
+    public abstract class Boundary2DMaxX : Boundary2D { }
+}
