+++ FDTD/Space2D/Boundaries/Boundary2DMaxY.cs
@@ -1,0 +1,4 @@
+﻿namespace FDTD.Space2D.Boundaries
+{
+    public abstract class Boundary2DMaxY : Boundary2D { }
+}
