+++ FDTD/Space2D/Boundaries/Boundary2DMinX.cs
@@ -1,0 +1,4 @@
+﻿namespace FDTD.Space2D.Boundaries
+{
+    public abstract class Boundary2DMinX : Boundary2D { }
+}
