+++ FDTD/Space2D/Boundaries/Boundary2DMinY.cs
@@ -1,0 +1,4 @@
+﻿namespace FDTD.Space2D.Boundaries
+{
+    public abstract class Boundary2DMinY : Boundary2D { }
+}
