+++ Tests/ConsoleTest/Solver1DTest.cs
@@ -1,6 +1,8 @@
 ﻿using System.IO;
 using FDTD;
 using FDTD.Space1D;
+using FDTD.Space1D.Boundaries.ABC;
+using FDTD.Space1D.Sources;
 using static System.Math;
 
 namespace ConsoleTest
